+++ internal/buildconfig/docker_env_test.go
@@ -1,4 +1,4 @@
-package buildspec
+package buildconfig
 
 import (
 	"reflect"
