+++ internal/buildconfig/params.go
@@ -1,4 +1,4 @@
-package buildspec
+package buildconfig
 
 import (
 	"flag"
