+++ internal/buildconfig/params_test.go
@@ -1,4 +1,4 @@
-package buildspec
+package buildconfig
 
 // TODO fix
 // import (
