+++ internal/buildconfig/testdata/bad-checksums.txt
