+++ internal/buildconfig/testdata/checksums.txt
